requirements/dev.txt
-r base.txt
django-debug-toolbar==1.2.2
django_extensions==1.4.9
